Added WaypointController
