+++ TeamCode/src/main/java/developing/Tracker.java
@@ -1,0 +1,24 @@
+package developing;
+
+import com.qualcomm.robotcore.util.ElapsedTime;
+
+import java.util.ArrayList;
+
+import globalfunctions.Storage;
+import globalfunctions.TimeData;
+
+public class Tracker {
+    public Storage storage = new Storage();
+    public ArrayList<double[]> track = new ArrayList<>();
+    public ArrayList<double[]> targets = new ArrayList<>();
+    public ArrayList<Double> trackTimes = new ArrayList<>();
+    public ArrayList<double[]> speeds = new ArrayList<>();
+    public ElapsedTime trackTime = new ElapsedTime();
+
+    public void saveData(){
+        storage.saveTimeData(new TimeData("Current", track, trackTimes));
+        storage.saveTimeData(new TimeData("Current2", speeds, trackTimes));
+//        storage.saveTimeData(new TimeData("Current3", poses, false));
+        storage.saveTimeData(new TimeData("Current4", targets, false));
+    }
+}
